behavioral_auth-20260228T031602Z-1-001/behavioral_auth/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
scikit-learn>=1.3
numpy>=1.24
